+++ controller/Changelog.md
@@ -1,6 +1,13 @@
 # [UNRELEASED]
 
+## Added
+
 - os: Add command for collecting system diagnostics
+- os: Local metric collection
+
+## Changed
+
+- driver: Upgrade driver to 2.8.0 for Flex V6 and device metadata support
 
 # [2026.1.0] - 2026-02-09